auth/requirements.txt
alembic==1.14.0
fastapi==0.115.5
passlib==1.7.4
psycopg[binary]==3.2.3
pydantic-settings==2.6.1
PyJWT==2.10.1
python-multipart==0.0.19
sqlmodel==0.0.22
uvicorn==0.32.1
